Trying to fix LauncherSQLiteHelper class which was broken after adding field activityName field to AppTable and ShortcutTable.
All methods which used to accept String packageName and then instantiates
an AppTable(id=null, packageName) inside it, now accepts AppTable directly.
Ditto for ShortcutTable.

I first discarded the thought of this interface change, fearing it would
introduce additional dependency and code coupling
(towards AppTable and ShortcutTable) to the classes like LauncherActivity.
Then I realised that LauncherSQLiteHelper already returns and accepts
List<AppTable> and List<ShortcutTable>, so this shouldn't be a big deal.

TODO: Figure out is it bad to have coupling between the rest of the code
and AppTable and ShortcutTable classes.
TODO: Figure out is it necessary to have a separate class .models.AppDetail
which is designed to contain the same information
already contained by AppTable class.