Fixed 'Add to Homescreen' option.
Fixed a bug in calculating the drag offset, which determined
when 'Add to Homescreen' option would popup.

`DragEvent` now contains `activityName` in addition to `packageName`
in its `clipData`.

Updated all methods which is concerned with adding shortcuts to pass
`AppTable` and `ShortcutTable` instead of just `packageName`.

`LauncherSQLiteHelper` now checks both `activityName` and `packageName`
to determine `canAddShortcut`.