Homescreen shortcuts now store the Intent extras passed in while creating shortcuts.
Thanks to @stanipintjuk for the comments and documentation,
they helped me to find the solution quickly.

`intent.toUri` is now stored in `ShortcutTable` and `AppDetail` classes.
If not null, it will be used to reconstruct the intent to be passed into
 `startActivity` using `Intent.parseUri`.

The unique index on activityName was removed from ShortcutTable and AppTable
for two reasons.
One, Multiple shortcuts can have same activityName but different intentUris.
Two, Multiple apps can have same activityName but different package name.
For example Microsoft Office apps, like Excel, Word etc.

canAddShortcut now checks for uniqueness taking into account intentUri
when it is not null.