Fixed a minor bug.
canAddShortcut used to return false when called:
    `canAddShortcut(new ShortcutTable(null,packageName,activityName,null/*intentUri*/);`
if an entity with non null intentUri exists in the database.
This is undesirable because, it may be needed to have a shortcut to filemanager,
and a shortcut to a file provided by that filemanager.
In this case the shortcut to filemanager was impossible to be added
if a file shortcut already exists.