Fixed: Pressing home key doesn't bring back launcher
When launching many apps like SoundCloud, Materialistic for example,
it kind of replaces the launcher.
When I press the recent apps button, the 'replaced' app is not shown
there. When I select another app from recents list and press home
button, it takes me back to the 'replaced' app, ie SoundCloud or
Materialistic.

Looks like Android confuses the activity we launch as a part of the
launcher app, instead of considering as a separate app.